Towards hot tls reload
